refactor(infra): extract diagnostics module to share origin helpers with future error pipeline
