+++ src/raitap/utils/__init__.py
@@ -2,7 +2,8 @@
 
 from __future__ import annotations
 
+from raitap.utils.diagnostics import Diagnostic
 from raitap.utils.serialization import to_json_serialisable
-from raitap.utils.warnings import WarningOrigin, suppress_warning
+from raitap.utils.warnings import suppress_warning
 
-__all__ = ["WarningOrigin", "suppress_warning", "to_json_serialisable"]
+__all__ = ["Diagnostic", "suppress_warning", "to_json_serialisable"]
